camera demo separate gaps implementation
