+++ demos/camera_demo/src/main-monolith.cpp
@@ -1,0 +1,162 @@
+/*
+ * This work was authored by Two Six Labs, LLC and is sponsored by a subcontract
+ * agreement with Galois, Inc.  This material is based upon work supported by
+ * the Defense Advanced Research Projects Agency (DARPA) under Contract No.
+ * HR0011-19-C-0103.
+ *
+ * The Government has unlimited rights to use, modify, reproduce, release,
+ * perform, display, or disclose computer software or computer software
+ * documentation marked with this legend. Any reproduction of technical data,
+ * computer software, or portions thereof marked with this legend must also
+ * reproduce this marking.
+ *
+ * Copyright 2020 Two Six Labs, LLC.  All rights reserved.
+ */
+
+#include <atomic>
+#include <chrono>
+#include <functional>
+#include <iostream>
+#include <memory>
+#include <string>
+#include <thread>
+#include <vector>
+
+#include <signal.h>
+#include <sys/signalfd.h>
+#include <unistd.h>
+
+#include "orientationinputcreator.hpp"
+#include "orientationoutputcreator.hpp"
+#include "frameprocessorcreator.hpp"
+#include "videosourcecreator.hpp"
+#include "imageconvert.hpp"
+#include "colortracking.hpp"
+#include "options.hpp"
+#include "optionsparser.hpp"
+
+static std::atomic<bool> interrupted(false);
+
+static int waitInterrupt(void* arg) {
+    (void) arg;
+    sigset_t set;
+    struct signalfd_siginfo unused;
+    int fd;
+
+    sigemptyset(&set);
+    sigaddset(&set, SIGINT);
+    pthread_sigmask(SIG_UNBLOCK, &set, NULL);
+
+    fd = signalfd(-1, &set, 0);
+    read(fd, &unused, sizeof(unused));
+
+    interrupted = true;
+    return 0;
+}
+
+int main(int argc, char *argv[])
+{
+    int rv;
+    Options options;
+    sigset_t set;
+    std::thread *signalThread;
+    std::vector<std::shared_ptr<FrameProcessor>> frameProcessors;
+
+    // block SIGINT for all threads except for the signalThread
+    sigemptyset(&set);
+    sigaddset(&set, SIGINT);
+    pthread_sigmask(SIG_BLOCK, &set, NULL);
+
+    parseArgs(argc, argv, &options);
+
+    OrientationOutput *orientationOutput;
+    std::vector<std::shared_ptr<OrientationInput>> orientationInputs;
+    std::shared_ptr<ColorTracking> colorTracking = nullptr;
+
+    orientationOutput = OrientationOutputCreator::get(options);
+
+    CameraOrientationCallbacks angPosCallbacks = orientationOutput->getCallbacks();
+    if (options.mImageTracking) {
+        colorTracking = std::make_shared<ColorTracking>(options, angPosCallbacks);
+        orientationInputs.push_back(colorTracking);
+    }
+
+    if (options.mInputKeyboard) {
+        std::shared_ptr<OrientationInput> io =
+            std::shared_ptr<OrientationInput>(OrientationInputCreator::get(Keyboard, options, angPosCallbacks));
+        orientationInputs.push_back(io);
+    }
+
+    if (options.mInputFreespace) {
+        std::shared_ptr<OrientationInput> io =
+            std::shared_ptr<OrientationInput>(OrientationInputCreator::get(Freespace, options, angPosCallbacks));
+        orientationInputs.push_back(io);
+    }
+
+    if (options.mFilesystemProcessor) {
+        FrameProcessorCreator::add(Filesystem, frameProcessors, options, angPosCallbacks);
+    }
+
+    if (options.mXWinProcessor) {
+        FrameProcessorCreator::add(XWindows, frameProcessors, options, angPosCallbacks);
+    }
+
+    if (options.mH264Encoder) {
+        FrameProcessorCreator::add(H264Stream, frameProcessors, options, angPosCallbacks);
+    }
+
+    if (options.mImageTracking) {
+        // Add color tracking to the end of frame processors.
+        // Take advantage of any RGB conversion in previous
+        // frame processors.
+        frameProcessors.push_back(colorTracking);
+    }
+
+    VideoSource *videoSource = VideoSourceCreator::create(options.mVideoInputType, frameProcessors, options);
+
+    rv = orientationOutput->init();
+    if (rv != 0)
+    {
+        return -1;
+    }
+
+    for (auto orientationInput : orientationInputs) {
+        rv = orientationInput->init();
+        if (rv != 0)
+        {
+            return -1;
+        }
+    }
+
+    for (auto frameProcessor : frameProcessors) {
+        rv = frameProcessor->init();
+        if (rv != 0)
+        {
+            return -1;
+        }
+    }
+
+    rv = videoSource->init();
+    if (rv != 0)
+    {
+        videoSource->term();
+        return -1;
+    }
+
+    signalThread = new std::thread(waitInterrupt, nullptr);
+
+    while (!interrupted)
+    {
+        std::this_thread::sleep_for(std::chrono::seconds(1));
+    }
+
+    signalThread->join();
+
+    delete signalThread;
+    delete videoSource;
+    frameProcessors.clear();
+    orientationInputs.clear();
+    delete orientationOutput;
+
+    return 0;
+}
